backend/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
openai>=1.10.0
qdrant-client>=1.8.0
python-dotenv>=1.0.0
tiktoken>=0.5.2
pydantic-settings>=2.1.0
python-multipart>=0.0.6